fix: Correctly remove snapshots from process state results
This PR ensures that process snapshot messages are removed from compute results in the `~process@1.0` device. Emitting these binaries directly is against the intended behavior of the device, and led to downstream encoding issues. These manifested occasionally as HTTP hangups, due to attempts to send unencoded values in message headers.