feat: Invariant-based testing framework for AO-Core and HyperBEAM.
A testing framework for AO-Core devices and HyperBEAM components built
upon the principles of property-based testing. Rather than testing specific
input and output pairs, `hb_invariant' allows us to instead focus on
defining invariant properties that should hold true for all valid inputs.
`hb_invariant' gives us tools to quickly and easily generate random inputs
(states, requests, node messages, etc.) to our components and then test that
the stated properties hold true for each of them.

## Execution Types.

Executions can come in a variety of forms:

- AO-Core device key relationships: Allowing us to define properties
  that should hold true for all `Base`, `Request`, node messages, and their
  corresponding `Result` messages.
- AO-Core device state machines: Allowing us to generate random initial
  states and sequences of requests, ensuring that a set of properties hold
  true at all times.
- Comparisons between two AO-Core device state machines: As above, except
  allowing us to define two generators for initial states, such that the
  functionality of one device can easily be compared to another. Properties
  in such tests receive not only the 'pre' and 'post' states for the primary
  state machine, but also the corresponding values for the reference machine.
- Direct Erlang function executions: Possible in each of the above cases,
  `hb_invariant' allows us to compute Erlang functions rather than AO-Core
  (`ao(Base, Req, Opts)') invocations, if preferred. This allows us to utilize
  `hb_invariant' to test HyperBEAM itself, as well as devices resident
  inside it.

## Execution Flow.

There are two primary invocation methods for `hb_invariant': `forall/1' and
`state_machine/1'. Because the state machine is sufficiently general to cover
all cases, under-the-hood `forall' is simply a wrapper around `state_machine'
that sets the length of the request sequence to `1'. A consequence of this
is that all invocations are able to utilize the full set of parameters to
control the execution.

The state machine executor always takes a `Specification' message as an
argument, and operates in a series of stages:

```
1. Specification normalization: All non-mandatory fields are filled in with
   default values, internal state keys are initialized in the `Spec', and
   initial seeding of the PRNG (`rand' module) is performed.
2. Repeat for each of the `Spec/runs' of the state machine:
2.1* Generate a node message (`Opts').
2.2* Generate an initial state (`Base' message) for the execution.
2.3* Generate an initial model state (`Model' message) for the execution, if
     applicable.
2.4. For each element of request sequence `Spec/length`:
2.4.1* Generate a request message (`Request' message) for the execution.
2.4.2* Execute the request message against the current state (and model state,
       if applicable), resulting in a `Result' message.
2.4.3. For each of the `Spec/properties':
2.4.3.1. Attempt to invoke the property function with the prior state(s), request,
       result(s), and options.
2.4.3.2. If the property function returns `true', continue to the next property.
2.4.3.3. If the property function returns `false', fail and return details of
         the executed sequence and error encountered.
2.4.3.4. If the property function lacks a function clause matching the call
         the failure is ignored. This allows callers to easily define which
         states are relevant for a given property simply with patterns and
         guards in the function head.
2.4.4. Apply `Spec/next' to the state and model state, if applicable, resulting
       in a new state and model state. If no `next' function is provided, the
       result of the request stage is used in the next iteration of the loop.
3. Return `ok' if all properties were enforced successfully, otherwise return
   details of the executed sequence and the error encountered.
'''
`*' markers above indicate that prior to the execution of a stage, the `rand'
module's PRNG is seeded with a value derived from the global seed (either
provided or generated at start time), the run number, the current request
count, and the current stage. This allows for reproducibility of the execution
sequences. See `Controlled Randomness' below for more details.

## Generators.

`hb_invariant' supports a number of different types of `generators', utilized
to derive each input in execution sequences. Supported generator forms are
as follows:

- Lists: Lists of generators of other forms, from which one one member is
  randomly selected and executed as if it was provided directly.
- Functions: Arbitrary Erlang functions, invoked with a specific set of
  arguments depending on the type of generator and the context.
- Explicit values: A simple constant value or message, used without execution.

Generators of these forms may be provided by the caller for each of the
keys listed below. Their names and function signatures are as follows:
- `opts(Spec)': A generator for the node message to use for a `run' of the
  state machine.
- `Spec/states': A generator for initial (`Base') states, executed per `run'.
- `Spec/models': A generator for initial _model_ states, executed per `run'.
- `Spec/requests': Generator of `Request's in the state transformation sequence.

In all cases aside `Spec/requests', the generator is optional, using a
default value if not provided. Without a `requests' generator, no sensible
state transformation sequence can be generated. Subsequently, execution is
aborted with an error.

## Controlled Randomness.

In order to assist in the creation of generators and properties for
`hb_invariant', a number of helper functions are provided to quickly and
easily generate random inputs of a given type. `hb_invariant' seeds Erlang's
`rand' module with a value derived from a provided global seed, or a unique
value per invocation of the state machine executor. In event of errors, the
initial global seed is provided to the user such that issues that arose may
be reproduced.

Value generators for the following types are provided:
- `int/0': Generate a random integer between 0 and the maximum 'small'
  (non-bignum) integer value.
- `int/1': Generate a random integer between 0 and the given maximum value.
- `int/2': Generate a random integer between the given values.
- `float/0': Generate a random float between 0 and the maximum float value.
- `float/1': Generate a random float between 0 and the given maximum value.
- `string/0': Generate a random string of a given length.
- `string/1': Generate a random string of a given length.
- `string/4': Generate a random string of a given length, with a give
  minimum and maximum character values, and a list of forbidden characters.